+++ src/TxHistory/TxHistoryNavigator.tsx
@@ -4,10 +4,10 @@ import {defineMessages, useIntl} from 'react-intl'
 import {StyleSheet, Text, TouchableOpacity, TouchableOpacityProps} from 'react-native'
 
 import {Boundary, Icon} from '../components'
-import {SelectTokenFromListScreenNavigator} from '../features/Send/common/SelectTokenFromListScreenNavigator'
 import {SendProvider} from '../features/Send/common/SendContext'
 import {ConfirmTxScreen} from '../features/Send/useCases/ConfirmTx/ConfirmTxScreen'
 import {ListAmountsToSendScreen} from '../features/Send/useCases/ListAmountsToSend'
+import {SelectTokenFromListScreen} from '../features/Send/useCases/ListAmountsToSend/AddToken/SelectTokenFromListScreen'
 import {EditAmountScreen} from '../features/Send/useCases/ListAmountsToSend/EditAmount/EditAmountScreen'
 import {ReadQRCodeScreen} from '../features/Send/useCases/StartMultiTokenTx/InputReceiver/ReadQRCodeScreen'
 import {StartMultiTokenTxScreen} from '../features/Send/useCases/StartMultiTokenTx/StartMultiTokenTxScreen'
@@ -18,6 +18,7 @@ import {
   useWalletNavigation,
 } from '../navigation'
 import {ReceiveScreen} from '../Receive/ReceiveScreen'
+import {useSearchHeaderOptions} from '../Search/SearchHeader'
 import {useSelectedWallet} from '../SelectedWallet'
 import {COLORS} from '../theme'
 import {useWalletName} from '../yoroi-wallets'
@@ -34,6 +35,11 @@ export const TxHistoryNavigator = () => {
   const [modalInfoState, setModalInfoState] = React.useState(false)
   const showModalInfo = () => setModalInfoState(true)
   const hideModalInfo = () => setModalInfoState(false)
+
+  const searchHeaderOptions = useSearchHeaderOptions({
+    placeHolderText: strings.searchPlaceholder,
+    title: strings.searchTitle,
+  })
 
   return (
     <SendProvider key={wallet.id}>
@@ -90,11 +96,16 @@ export const TxHistoryNavigator = () => {
           )}
         </Stack.Screen>
 
-        <Stack.Screen //
-          name="send"
-          component={SelectTokenFromListScreenNavigator}
-          options={{headerShown: false}}
-        />
+        <Stack.Screen
+          name="send-select-token-from-list"
+          options={{...defaultStackNavigationOptionsV2, ...searchHeaderOptions}}
+        >
+          {() => (
+            <Boundary>
+              <SelectTokenFromListScreen />
+            </Boundary>
+          )}
+        </Stack.Screen>
 
         <Stack.Screen //
           name="send-list-amounts-to-send"
@@ -186,6 +197,14 @@ const messages = defineMessages({
       'To protect your privacy, new addresses are ' +
       'generated automatically once you use them.',
   },
+  searchTitle: {
+    id: 'components.send.sendscreen.searchTitle',
+    defaultMessage: '!!!Select Assets',
+  },
+  searchPlaceholder: {
+    id: 'components.send.sendscreen.searchPlaceholder',
+    defaultMessage: '!!!Search Assets',
+  },
 })
 
 const useStrings = () => {
@@ -200,6 +219,8 @@ const useStrings = () => {
     receiveInfoText: intl.formatMessage(messages.receiveInfoText),
     editAmountTitle: intl.formatMessage(messages.editAmountTitle),
     listAmountsToSendTitle: intl.formatMessage(messages.listAmountsToSendTitle),
+    searchTitle: intl.formatMessage(messages.searchTitle),
+    searchPlaceholder: intl.formatMessage(messages.searchPlaceholder),
   }
 }
