+++ src/features/Send/useCases/ListAmountsToSend/ListAmountsToSendScreen.tsx
@@ -8,7 +8,7 @@ import {useQuery, UseQueryOptions} from 'react-query'
 import {Boundary, Button, Spacer} from '../../../../components'
 import {AmountItem} from '../../../../components/AmountItem/AmountItem'
 import globalMessages from '../../../../i18n/global-messages'
-import {TxHistoryRouteNavigation, useWalletNavigation} from '../../../../navigation'
+import {TxHistoryRouteNavigation} from '../../../../navigation'
 import {useSelectedWallet} from '../../../../SelectedWallet'
 import {COLORS} from '../../../../theme'
 import {sortTokenInfos} from '../../../../utils'
@@ -173,10 +173,9 @@ const messages = defineMessages({
 
 const useNavigateTo = () => {
   const navigation = useNavigation<TxHistoryRouteNavigation>()
-  const {navigateSendSelectAssetFromList} = useWalletNavigation()
 
   return {
-    addToken: () => navigateSendSelectAssetFromList(),
+    addToken: () => navigation.navigate('send-select-token-from-list'),
     editAmount: () => navigation.navigate('send-edit-amount'),
     confirmTx: () => navigation.navigate('send-confirm-tx'),
   }
