+++ src/features/Send/useCases/StartMultiTokenTx/StartMultiTokenTxScreen.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import {KeyboardAvoidingView, Platform, ScrollView, StyleSheet, View, ViewProps} from 'react-native'
 
 import {Button, Spacer} from '../../../../components'
-import {TxHistoryRouteNavigation, useWalletNavigation} from '../../../../navigation'
+import {TxHistoryRouteNavigation} from '../../../../navigation'
 import {useSelectedWallet} from '../../../../SelectedWallet'
 import {COLORS} from '../../../../theme'
 import {useHasPendingTx, useIsOnline} from '../../../../yoroi-wallets/hooks'
@@ -43,7 +43,7 @@ export const StartMultiTokenTxScreen = () => {
 
   const onNext = () => {
     navigateTo.selectedTokens()
-    if (shouldOpenSelectToken) navigateTo.selectToken()
+    if (shouldOpenSelectToken) navigateTo.addToken()
   }
 
   return (
@@ -89,11 +89,10 @@ const Actions = ({style, ...props}: ViewProps) => <View style={[styles.actions, 
 
 const useNavigateTo = () => {
   const navigation = useNavigation<TxHistoryRouteNavigation>()
-  const {navigateSendSelectAssetFromList} = useWalletNavigation()
 
   return {
     selectedTokens: () => navigation.navigate('send-list-amounts-to-send'),
-    selectToken: () => navigateSendSelectAssetFromList(),
+    addToken: () => navigation.navigate('send-select-token-from-list'),
   }
 }
